feat: remove dap capabilities
After having experimented with debuggers for a while, I realised that
they aren't best fit for my personal workflow. Setting up a debugger can
be a bit tedious and requires quite a few dependencies. Other static
analysis tools help fill in some of the void left without debuggers, but
ultimately it is not as big of a loss.